requirements.txt
celery==5.3.5
elasticsearch==8.10.1
fastapi==0.104.1
gunicorn==21.2.0
httpx==0.25.1
mongoengine==0.27.0
pymongo==4.6.0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
uvicorn==0.24.0.post1
beautifulsoup4
flower
colorlog